Add detail product admin, fix bug

+++ Modules/Admin/Http/Controllers/ProductController.php
@@ -43,8 +43,9 @@ class ProductController extends Controller
     public function renderPropertyForm(Request $request) {
         if($request->ajax()) {
             $categoryId = $request->get('category_id');
+            $oldDatas = $request->get('old_data');
             $propertyTypes = $this->productService->getPropertyTypeByCategoryId($categoryId);
-            $propertyForm = view('admin::products.property-form', compact('propertyTypes'))->render();
+            $propertyForm = view('admin::products.property-form', compact('propertyTypes', 'oldDatas'))->render();
             return response()->json(compact('propertyForm'));
         } else {
             return response()->json([]);
@@ -65,8 +66,15 @@ class ProductController extends Controller
 
     public function delete(Request $request) {
         $ids = $request->get('ids');
+        if(empty($ids)) return back();
         $this->productService->deleteMultiProduct($ids);
 
         return redirect()->route('admin.product.list');
+    }
+
+    public function detail(Request $request, $id) {
+        $product = $this->productService->getProductById($id);
+
+        return view('admin::products.detail', compact('product'));
     }
 }
